HDDS-12420. Move FinalizeUpgradeCommandUtil to hdds-common
